Ran some roc tests with my histogram and chi square. Not so great, but will look for ways to improve

+++ experiments/local_binary_patterns/uniform_lbp.m
@@ -30,7 +30,7 @@ if isempty(lutable)
     end
 end
 
-uniform_code = lutable(code+1);
+uniform_code = lutable( uint16(code)+1 );
 
 end
